keep only message.(Un)Marshal methods
